requirements.txt
asgiref==3.4.1
boto3==1.20.23
botocore==1.23.24
Django==3.2.9
django-appconf==1.0.5
django-crispy-forms==1.13.0
django-imagekit==4.1.0
django-resized==0.3.11
django-storages==1.12.3
gunicorn==20.1.0
jmespath==0.10.0
pilkit==2.0
Pillow==8.4.0
python-dateutil==2.8.2
pytz==2021.3
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
whitenoise==5.3.0
